Moved XmlExceptions to separate header
